Implementación de comparación con density-based

+++ lib/utils.hpp
@@ -9,8 +9,8 @@ using namespace std;
 struct astar_node {
     int v; // Current node
     int p; // Node's parent
-    int f; // Estimated cost from the source to the goal
-    int g; // Accumulated cost from the source to v
+    float f; // Estimated cost from the source to the goal
+    float g; // Accumulated cost from the source to v
 };
 
 bool operator<(const astar_node& a, const astar_node& b);
